Update *.md files, Semaphore CI, Makefile and test to recent standard
This commit applies format updates to all common Markdown files
to conform to the most recent standard. The Semaphore CI config
and Makefile are updated as well.

Furthermore, unnecessary link check flags are removed.

Finally, the terratest method is set to InitAndApplyAndIdempotent
and the terratest release number is set to the version containing this
method.

Things to check:
- The semaphore config still contains a `auto_cancel` block not found elsewhere.
- The `version.tf` does not follow the ">= 0.12.20, < 0.14.0" syntax as per the module repository. Several other recent repositories do not as well so without being sure what the preferred way is, I left it unchanged.

Things not yet covered:
- No integration tests yet